# performance improvements span based serializers