--- a/Source/MQTTnet/Adapter/ReceivedMqttPacket.cs
+++ b/Source/MQTTnet/Adapter/ReceivedMqttPacket.cs
@@ -4,16 +4,16 @@ namespace MQTTnet.Adapter
 {
     public class ReceivedMqttPacket
     {
-        public ReceivedMqttPacket(byte fixedHeader, MqttPacketBodyReader body, int totalLength)
+        public ReceivedMqttPacket(byte fixedHeader, IMqttPacketBodyReader body, int totalLength)
         {
             FixedHeader = fixedHeader;
             Body = body;
             TotalLength = totalLength;
         }
 
-        public byte FixedHeader { get; }
+        public byte FixedHeader { get; set; }
 
-        public MqttPacketBodyReader Body { get; }
+        public IMqttPacketBodyReader Body { get; }
 
         public int TotalLength { get; }
     }