… Source/MQTTnet/Formatter/IMqttPacketWriter.cs
… Source/MQTTnet/Formatter/MqttPacketBodyReader.cs
… Source/MQTTnet/Formatter/MqttPacketFormatterAdapter.cs
… Source/MQTTnet/Formatter/V3/MqttV310PacketFormatter.cs
… Source/MQTTnet/Formatter/V3/MqttV311PacketFormatter.cs
… Source/MQTTnet/Formatter/V5/MqttV500PacketDecoder.cs
… Source/MQTTnet/Formatter/V5/MqttV500PacketEncoder.cs
… Source/MQTTnet/Formatter/V5/MqttV500PacketFormatter.cs
… Source/MQTTnet/Formatter/V5/MqttV500PropertiesReader.cs
… Source/MQTTnet/Formatter/V5/MqttV500PropertiesWriter.cs
… Tests/MQTTnet.AspNetCore.Tests/MQTTnet.AspNetCore.Tests.csproj
… Tests/MQTTnet.AspNetCore.Tests/MqttPacketSerializerTests.cs
… Tests/MQTTnet.AspNetCore.Tests/ReaderExtensionsTest.cs
… Tests/MQTTnet.Benchmarks/Configurations/AllowNonOptimized.cs
… Tests/MQTTnet.Benchmarks/Configurations/BaseConfig.cs
… Tests/MQTTnet.Benchmarks/Configurations/RuntimeCompareConfig.cs
… Tests/MQTTnet.Benchmarks/MQTTnet.Benchmarks.csproj
… Tests/MQTTnet.Benchmarks/Program.cs
… Tests/MQTTnet.Core.Tests/MqttPacketSerializer_Tests.cs