I cannot compare with before because the current dev branch is not able to run the benchmark.

see commit "fixed another issue", seems the mqtt v5 changes introduced an issue in the aspnet part of the codebase.


this is what it looks like after my changes:

|              Method | Runtime |     Toolchain |     Mean |     Error |    StdDev |     Gen 0 |    Gen 1 |   Gen 2 | Allocated |
|-------------------- |-------- |-------------- |---------:|----------:|----------:|----------:|---------:|--------:|----------:|
| Send_10000_Messages |     Clr |       Default | 25.30 ms | 0.4913 ms | 0.6213 ms | 2562.5000 |  93.7500 | 31.2500 |   10.7 MB |
| Send_10000_Messages |    Core | .NET Core 2.1 | 12.19 ms | 0.3646 ms | 1.0225 ms | 1718.7500 | 546.8750 | 31.2500 |   1.45 MB |

vs 2019 includes an improved profiler it seems with it I can spot many more allocations per message I would like to get rid of. most of them require changes of apis so I left them like they are and just focused on reducing the number of byte[] allocations.

ValueTask instead of Task was one option I saw benifit in my first attempt but to keep the scope of the change smaller I left that like it is. 

![image](https://user-images.githubusercontent.com/6747376/56138361-0def5580-5f97-11e9-8d41-5468b24d8847.png)

@chkr1011 pls do a review asap. I dont want the branch to get stale again.