@chkr1011 2019-04-16
Hi, 
ValueTasks look interesting. Didn't know that this exists...
What do you plan for API changes? I want to release 3.0 soon and keep the API stable for a while.