@JanEggers 2019-04-16
> What do you plan for API changes? I want to release 3.0 soon and keep the API stable for a while.

thought so, that is why I didnt change em. basic idea is to get as close as possible to zero allocations / packet. I understand that we need some of the stuff on the screenshot but I wonder if we can get rid of the collections and hashsets. In general I want to speed up processing of messages I have the broker running on an embedded device and currently im at 50% i need to reduce that as much as possible.

> What do you think about a .net standard 2.0 library version?
I dont know if we would gain anything from that. We need AspNetCore (Kestrel) to get the pipelines based tcp / websocket server. I dont see a benifit in implementing that ourself. What is your motivation? 

> at least for users which are able to use latest and greatest versions.

Im not sure how microsoft / netstandard will evolve. I raised an Issue about a new features in System.Memory.dll (about a SequenceReader) but they said they will not release it as seperate nuget only part of netcore 3.0 and that will not be compatible with netfx / netstandard 2.1. for me thats a big downside as I thought with dotnet core the monolith framework was properly broken down, but right now it looks like the same mess as netfx, just cross platform.