@chkr1011 2019-04-16
Yes I was hoping the pipelines etc. are available but if this is not possible then we don't need it. 

When it comes to allocations I also tried to initialize all properties (in mqtt5) with null by default but at some points it was not possible or too much effort. But we can try it again. The decoder can create them on demand (if user properties etc. are available). 

Did you saw the latest added question?

I added some TODOs in MqttConnectionContext. Do you have an idea how to fill them properly?